api: 0.9.1 deprecated damage sub-values and clarified docs
The API methods getImpactDamage() and setImpactDamage(damage) have been
deprecated. In the next major version the concept of damage types will
be removed from the ExplosiveImpactEvent, as the only damage values ever
passed to it will originate from the impact itself. A separate event
will be created and thrown as the explosion damages the player. This
event will support categories as it will have values for impact,
explosion, and combined damage.