Fix WorldGuard integration for WorldGuard 7+ (1.13+).
From the data on the metrics page, most servers are using 1.13+. As such, support for WorldGuard 6 has been completely dropped. The only way to add simultaneous support for both versions of the API would be to reflectively invoke the API calls. This wasn't worth it given the usage distribution.

ExplosiveElytras v0.11.0 will only support WorldGuard integration for v7+. This means that it will only be available on servers running 1.13+.

If you are still on an older version, WorldGuard integration will not be available. If you absolutely cannot upgrade (not recommended), you will need to use v0.10.2 for WorldGuard 6 support.